sbd/lab_3/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
WTForms==3.1.0
psycopg2-binary==2.9.7
pymongo==4.5.0
python-dotenv==1.0.0